+++ packages/gatsby-admin/src/pages/index.tsx
@@ -10,6 +10,11 @@ import {
   InputField,
   InputFieldControl,
   ButtonProps,
+  InputFieldLabel,
+  DropdownMenu,
+  DropdownMenuButton,
+  DropdownMenuItem,
+  DropdownMenuItems,
 } from "gatsby-interface"
 
 const SecondaryButton: React.FC<ButtonProps> = props => (
@@ -19,12 +24,6 @@ const SecondaryButton: React.FC<ButtonProps> = props => (
     sx={{
       paddingX: 6,
       paddingY: 4,
-      color: `whiteFade.80`,
-      border: `sixtywhite`,
-      "&:hover": {
-        color: `white`,
-        border: `white`,
-      },
     }}
     {...props}
   ></Button>
@@ -65,9 +64,7 @@ const InstallInput: React.FC<{ for: string }> = props => {
     >
       <InputField id={inputId}>
         <Flex gap={2} flexDirection="column">
-          <Text size="S" sx={{ color: `grey.40` }}>
-            <label htmlFor={inputId}>Install {props.for}:</label>
-          </Text>
+          <InputFieldLabel>Install {props.for}:</InputFieldLabel>
           <Flex gap={4} alignItems="center">
             <InputFieldControl
               placeholder={`gatsby-${props.for}-`}
@@ -75,15 +72,7 @@ const InstallInput: React.FC<{ for: string }> = props => {
               value={value}
               onChange={(e): void => setValue(e.target.value)}
               sx={{
-                backgroundColor: `background`,
-                borderColor: `grey.60`,
-                color: `white`,
                 width: `initial`,
-                "&:focus": {
-                  borderColor: `grey.40`,
-                  // TODO(@mxstbr): Fix this focus outline
-                  boxShadow: `none`,
-                },
               }}
             />
             <SecondaryButton
@@ -100,7 +89,13 @@ const InstallInput: React.FC<{ for: string }> = props => {
   )
 }
 
-const DestroyButton: React.FC<{ name: string }> = ({ name }) => {
+const SectionHeading: React.FC<HeadingProps> = props => (
+  <Heading as="h1" sx={{ fontWeight: `500`, fontSize: 5 }} {...props} />
+)
+
+const PluginCard: React.FC<{
+  plugin: { name: string; description?: string }
+}> = ({ plugin }) => {
   const [, deleteGatsbyPlugin] = useMutation(`
     mutation destroyGatsbyPlugin($name: String!) {
       destroyNpmPackage(npmPackage: {
@@ -122,46 +117,47 @@ const DestroyButton: React.FC<{ name: string }> = ({ name }) => {
   `)
 
   return (
-    <SecondaryButton
-      onClick={(evt): void => {
-        evt.preventDefault()
-        if (window.confirm(`Are you sure you want to uninstall ${name}?`)) {
-          deleteGatsbyPlugin({ name })
-        }
-      }}
+    <Flex
+      flexDirection="column"
+      gap={6}
+      sx={{ backgroundColor: `ui.background`, padding: 5, borderRadius: 2 }}
     >
-      Uninstall
-    </SecondaryButton>
+      <Flex justifyContent="space-between">
+        <Heading as="h2" sx={{ fontWeight: `500`, fontSize: 3 }}>
+          {plugin.name}
+        </Heading>
+        <DropdownMenu>
+          <DropdownMenuButton
+            aria-label="Actions"
+            sx={{
+              border: `none`,
+              background: `transparent`,
+              color: `text.secondary`,
+            }}
+          >
+            ···
+          </DropdownMenuButton>
+          <DropdownMenuItems>
+            <DropdownMenuItem
+              onSelect={(): void => {
+                if (
+                  window.confirm(`Are you sure you want to uninstall ${name}?`)
+                ) {
+                  deleteGatsbyPlugin({ name })
+                }
+              }}
+            >
+              Uninstall
+            </DropdownMenuItem>
+          </DropdownMenuItems>
+        </DropdownMenu>
+      </Flex>
+      <Text sx={{ color: `text.secondary` }}>
+        {plugin.description || <em>No description.</em>}
+      </Text>
+    </Flex>
   )
 }
-
-const SectionHeading: React.FC<HeadingProps> = props => (
-  <Heading
-    as="h1"
-    sx={{ color: `white`, fontWeight: `500`, fontSize: 5 }}
-    {...props}
-  />
-)
-
-const PluginCard: React.FC<{
-  plugin: { name: string; description?: string }
-}> = ({ plugin }) => (
-  <Flex
-    flexDirection="column"
-    gap={6}
-    sx={{ backgroundColor: `grey.80`, padding: 5, borderRadius: 2 }}
-  >
-    <Heading as="h2" sx={{ color: `white`, fontWeight: `500`, fontSize: 3 }}>
-      {plugin.name}
-    </Heading>
-    <Text sx={{ color: `grey.40` }}>
-      {plugin.description || <em>No description.</em>}
-    </Text>
-    <Flex justifyContent="flex-end" sx={{ width: `100%` }}>
-      <DestroyButton name={plugin.name} />
-    </Flex>
-  </Flex>
-)
 
 const Index: React.FC<{}> = () => {
   const [{ data, fetching, error }] = useQuery({