+++ packages/gatsby-cli/src/index.ts
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 
-import "@babel/polyfill"
 import os from "os"
 import semver from "semver"
 import util from "util"
@@ -46,7 +45,7 @@ if (semver.prerelease(version)) {
   report.warn(
     report.stripIndent(`
     You are currently using a prerelease version of Node (${version}), which is not supported.
-    You can use this for testing, but we do not recommend it in production. 
+    You can use this for testing, but we do not recommend it in production.
     Before reporting any bugs, please test with a supported version of Node (>=${MIN_NODE_VERSION}).`)
   )
 }