+++ packages/gatsby-plugin-google-analytics/src/gatsby-ssr.js
@@ -13,6 +13,7 @@ const knownOptions = {
     legacyCookieDomain: `string`,
     legacyHistoryImport: `boolean`,
     allowLinker: `boolean`,
+    storage: `string`,
   },
   general: {
     allowAdFeatures: `boolean`,
@@ -34,8 +35,13 @@ export const onRenderBody = (
   // Lighthouse recommends pre-connecting to google analytics
   setHeadComponents([
     <link
-      rel="preconnect dns-prefetch"
+      rel="preconnect"
       key="preconnect-google-analytics"
+      href="https://www.google-analytics.com"
+    />,
+    <link
+      rel="dns-prefetch"
+      key="dns-prefetch-google-analytics"
       href="https://www.google-analytics.com"
     />,
   ])