+++ packages/gatsby/src/commands/types.ts
@@ -3,8 +3,13 @@ import { Store, AnyAction } from "redux"
 import { IGatsbyState } from "../redux/types"
 
 export interface ICert {
-  key: string
-  cert: string
+  key: Buffer
+  cert: Buffer
+}
+
+export interface IDebugInfo {
+  port: number
+  break: boolean
 }
 
 export interface IProgram {
@@ -23,6 +28,8 @@ export interface IProgram {
   https?: boolean
   sitePackageJson: PackageJson
   ssl?: ICert
+  inspect?: number
+  inspectBrk?: number
   graphqlTracing?: boolean
   setStore?: (store: Store<IGatsbyState, AnyAction>) => void
 }