+++ packages/gatsby/src/services/start-webpack-server.ts
@@ -21,7 +21,7 @@ import {
   markWebpackStatusAsDone,
 } from "../utils/webpack-status"
 import { enqueueFlush } from "../utils/page-data"
-import { mapTemplatesToStaticQueryHashes } from "../utils/map-pages-to-static-query-hashes"
+import mapTemplatesToStaticQueryHashes from "../utils/map-templates-to-static-query-hashes"
 
 export async function startWebpackServer({
   program,
@@ -42,17 +42,25 @@ export async function startWebpackServer({
   )
 
   compiler.hooks.invalid.tap(`log compiling`, function () {
-    markWebpackStatusAsPending()
+    if (!webpackActivity) {
+      // mark webpack as pending if we are not in the middle of compilation already
+      // when input is invalidated during compilation, webpack will automatically
+      // run another compilation round before triggering `done` event
+      report.pendingActivity({ id: `webpack-develop` })
+      markWebpackStatusAsPending()
+    }
   })
 
   compiler.hooks.watchRun.tapAsync(`log compiling`, function (_, done) {
-    if (webpackActivity) {
-      webpackActivity.end()
+    if (!webpackActivity) {
+      // there can be multiple `watchRun` events before receiving single `done` event
+      // webpack will not emit assets or `done` event until all pending invalidated
+      // inputs were compiled
+      webpackActivity = report.activityTimer(`Re-building development bundle`, {
+        id: `webpack-develop`,
+      })
+      webpackActivity.start()
     }
-    webpackActivity = report.activityTimer(`Re-building development bundle`, {
-      id: `webpack-develop`,
-    })
-    webpackActivity.start()
 
     done()
   })
@@ -125,7 +133,7 @@ export async function startWebpackServer({
             if (
               !isEqual(
                 state.staticQueriesByTemplate.get(componentPath),
-                staticQueryHashes.map(String)
+                staticQueryHashes
               )
             ) {
               store.dispatch({