+++ packages/gatsby/src/utils/page-data.ts
@@ -54,6 +54,10 @@ export async function removePageData(
   return Promise.resolve()
 }
 
+export function pageDataExists(publicDir: string, pagePath: string): boolean {
+  return fs.existsSync(getFilePath(publicDir, pagePath))
+}
+
 export async function writePageData(
   publicDir: string,
   {
@@ -141,7 +145,7 @@ export async function flush(): Promise<void> {
     // This is why we need this check
     if (page) {
       const staticQueryHashes =
-        staticQueriesByTemplate.get(page.componentPath)?.map(String) || []
+        staticQueriesByTemplate.get(page.componentPath) || []
 
       const result = await writePageData(
         path.join(program.directory, `public`),